+++ src/main.h
@@ -138,10 +138,11 @@ static const unsigned int MAX_BLOCKS_TO_ANNOUNCE = 12;
 static const int MAX_UNCONNECTING_HEADERS = 10;
 
 //Dynamic Constants
-static const CAmount STATIC_POW_REWARD = COIN * 1;
-static const CAmount BLOCKCHAIN_INIT_REWARD = COIN * 0;
-static const CAmount STATIC_DYNODE_PAYMENT = COIN * 0.382;
 static const CAmount MIN_TXOUT_AMOUNT = MIN_TX_FEE;
+static const CAmount BLOCKCHAIN_INIT_REWARD = COIN * 0;
+static const CAmount PHASE_1_POW_REWARD = COIN * 1;
+static const CAmount PHASE_1_DYNODE_PAYMENT = COIN * 0.382;
+static const CAmount PHASE_2_DYNODE_PAYMENT = COIN * 0.618;
 
 struct BlockHasher
 {