+++ src/reverselock.h
@@ -18,10 +18,12 @@ public:
 
     explicit reverse_lock(Lock& _lock) : lock(_lock) {
         _lock.unlock();
+        lock.swap(templock);
     }
 
     ~reverse_lock() {
-        lock.lock();
+        templock.lock();
+        templock.swap(lock);
     }
 
 private:
@@ -29,6 +31,7 @@ private:
     reverse_lock& operator=(reverse_lock const&);
 
     Lock& lock;
+    Lock templock;
 };
 
 #endif // DYNAMIC_REVERSELOCK_H