+++ src/wallet/wallet.h
@@ -43,6 +43,7 @@ extern bool fSendFreeTransactions;
 extern bool fLargeWorkForkFound;
 extern bool fLargeWorkInvalidChainFound;
 
+//! -keypool default
 static const unsigned int DEFAULT_KEYPOOL_SIZE = 2000;
 //! -paytxfee default
 static const CAmount DEFAULT_TRANSACTION_FEE = 0;